==> 00-gerekli-r-kutuphaneleri.R <==
if (!require(tidyverse)) {
  install.packages("tidyverse",
    dependencies = TRUE
  )
}
if (!require(ggforce)) {
  install.packages("ggforce",
    dependencies = TRUE
  )
}
if (!require(DataExplorer)) {
  install.packages("DataExplorer",
    dependencies = TRUE
  )
}
if (!require(doParallel)) {
  install.packages("doParallel",
    dependencies = TRUE
  )
}
if (!require(sf)) install.packages("sf", dependencies = TRUE)
if (!require(rnaturalearth)) {
  install.packages("rnaturalearth",
    dependencies = TRUE
  )
}
if (require(rnaturalearthdata)) {
  install.packages("rnaturalearthdata", dependencies = TRUE)
}
if (!require(ggthemes)) {
  install.packages("ggthemes",
    dependencies = TRUE
  )
}
if (!require(ggplot2)) {
  install.packages("ggplot2",
    dependencies = TRUE
  )
}
if (!require("ggcorrplot")) {
  install.packages("ggcorrplot",
    dependencies = TRUE
  )
}
if (!requireNamespace("devtools")) install.packages("devtools")
devtools::install_github("caijun/ggcorrplot2")
if (!require("factoextra")) {
  install.packages("factoextra",
    dependencies = TRUE
  )
}
if (!require("forcats")) {
  install.packages("forcats",
    dependencies = TRUE
  )
}
if (!require("DALEX")) {
  install.packages("DALEX",
    dependencies = TRUE
  )
}
if (!require("DALEXtra")) {
  install.packages("DALEXtra",
    dependencies = TRUE
  )
}
if (!require(ggfortify)) {
  install.packages("ggfortify",
    dependencies = TRUE
  )
}

if (!require(skimr)) {
  install.packages("skimr",
    dependencies = TRUE
  )
}
if (!require(reticulate)) {
  install.packages("reticulate",
    dependencies = TRUE
  )
}
if (!require("tseries")) {
  install.packages("tseries",
    dependencies = TRUE
  )
}
if (!require("psych")) {
  install.packages("psych",
    dependencies = TRUE
  )
}

if (!require("recipes")) {
  install.packages("recipes",
    dependencies = TRUE
  )
}
if (!require("hopkins")) {
  install.packages("hopkins",
    dependencies = TRUE
  )
}
if (!require(lmtest)) {
  install.packages("lmtest",
    dependencies = TRUE
  )
}
if (!require(car)) {
  install.packages("car",
    dependencies = TRUE
  )
}
if (!require(broom)) {
  install.packages("broom",
    dependencies = TRUE
  )
}
if (!require(caret)) {
  install.packages("caret",
    dependencies = TRUE
  )
}
if (!require(MLmetrics)) {
  install.packages("MLmetrics",
    dependencies = TRUE
  )
}
if (!require("coin")) {
  install.packages("coin",
    dependencies = TRUE
  )
}

if (!require("rcompanion")) {
  install.packages("rcompanion")
}

if (!require("NbClust")) {
  install.packages("NbClust",
    dependencies = TRUE
  )
}
if (!require("clValid")) {
  install.packages("clValid",
    dependencies = TRUE
  )
}
if (!require("cluster")) {
  install.packages("cluster",
    dependencies = TRUE
  )
}
if (!require("mclust")) {
  install.packages("mclust",
    dependencies = TRUE
  )
}

if (!require("h2o")) {
  install.packages("h2o",
    dependencies = TRUE
  )
}
tryCatch(
  expr = {
    install_miniconda()
  },
  error = function(e) {
  },
  warning = function(w) {
  },
  finally = {
  }
)
use_miniconda("r-reticulate")
paketler <- readLines("pyFiles/requirements.txt")
conda_install(packages = paketler[!grepl(paketler, pattern = "^#")])
